refactor(MapFilterSlider)
